Splits block.rs into separate files
